move requests pages to new folder
