+++ src/pages/dashboard/requests/pages/ManageRequestPage.tsx
@@ -1,13 +1,13 @@
 import { institutionTableData } from "@/lib/helpers";
-import PageTitle from "./components/PageTitle";
-import SeachFilter from "./components/SeachFilter";
+import PageTitle from "../components/PageTitle";
+import SeachFilter from "../components/SeachFilter";
 import LinkIcon from "@/components/custom/Icons/LinkIcon";
 import { Button } from "@/components/ui/button";
 import GoogleDoc from "@/components/custom/Icons/GoogleDoc";
 import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import Loader from "@/components/custom/Loader";
-import ManageRequests from "./components/ManageRequest";
+import ManageRequests from "../components/ManageRequest";
 import { useMediaQuery } from "react-responsive";
 import { X } from "lucide-react";
